[refactor] Moved methods to separated folder
fix #244
