+++ src/popper/utils/clockwise.js
@@ -1,4 +1,4 @@
-import placements from '../utils/placements';
+import placements from '../methods/placements';
 
 // Get rid of `auto` `auto-start` and `auto-end`
 const validPlacements = placements.slice(3);
